backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pytesseract==0.3.13
Pillow==11.0.0
openai==1.54.4
python-dotenv==1.0.1
pydantic==2.9.2
SQLAlchemy==2.0.36
PyMySQL==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
